requirements.txt
xbox360controller~=1.1.2
pyserial~=3.5
c1c0_movement~=0.1.0
grpcio~=1.41.0
